Addded USB re-enumeration tool
On all the ROG laptops, HID Keyboard device is built-in and cannot be unplugged. DriverKit for some reason doesn't match to already enumerated devices. So, using this tool we can re-enumerate the devices based on VID and PID. That simulates a reconnect and allows DriverKit to attach.

Once you install the system extension, from then on, it will load and attach automatically.

Signed-off-by: black.dragon74 <nickk.2974@gmail.com>