Fix erratic caps lock key
For some weird reason, the CAPS Lock key on the HID keyboard doesn't work properly. You can reproduce this bug by tapping the key rapidly. The underlying problem seems to be related with setting the keyboard LED state. If a report of setLED is processing, the system ignores all the subsequent reports. So, we induce an IOSleep of 80ms helping system to properly set LED state and fix the erratic CAPS Lock key behavior in the process.

This fix should not hurt systems without a keyboard caps lock LED. However, if such is the case this fix should be re-implemented in a better way (custom implementaion of `SetLED` and `parseLEDElement`)

Signed-off-by: black.dragon74 <nickk.2974@gmail.com>