Dispatch lux queue only for Asus' HID Usage Page
Some ROG HID keyboards like 0x1866 have multiple IOHIDInterface nubs exposed. In my opinion, the two non-asus nubs handle the normal keyboard keys and have a GenericDesktop(kHIDPage_GenericDesktop) usage page. The nub with Asus' usage page is responsible for handling special Fn key features. Attaching to all three nubs caused backlight inactivity timer issues as our `lastDispatchTime` var is static (meant to be) and multiple queues from different threads wrote to this value causing the value to be overwritten and inconsistent. In this patch, we don't dispatch any queues for non Asus HID pages. So we are assured that only a single thread will read and write to the variable and will hence fix this issue. See issue #4 for detailed discussion on this.

Signed-off-by: black.dragon74 <nickk.2974@gmail.com>